src/agents/page_info_extractor/requirements.txt
# Google Agent Development Kit (ADK)
google-adk

# Web crawling dependencies
httpx
selectolax
tldextract

# Configuration and data validation
pydantic
pydantic-settings
python-dotenv

# Google Generative AI (for genai.types)
google-genai

# Async support
asyncio-mqtt